# backend/openapi.yaml
openapi: 3.0.0
info:
  title: CollabSync API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /auth/register:
    post:
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "200":
          description: Successful registration with a token
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  auth_token:
                    type: string
  /auth/login:
    post:
      summary: Login a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Successful login with a token
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  auth_token:
                    type: string
  /users:
    get:
      summary: Search users
      operationId: searchUsers
      parameters:
        - in: query
          name: query
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
        - in: query
          name: offset
          schema:
            type: integer
            format: int32
        - in: query
          name: sort_by
          schema:
            type: string
            enum:
              - name
              - created_at
        - in: query
          name: sort_order
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
components:
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
        password_hash:
          type: string
        role:
          type: string
    CreateUserInput:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password_hash:
          type: string
        role:
          type: string
...